move nanoflann header to common.h
