+++ cpp/open3d/core/nns/NeighborSearchCommon.h
@@ -26,7 +26,23 @@
 
 #pragma once
 
+#include <memory>
+
 #include "open3d/utility/MiniVec.h"
+
+namespace nanoflann {
+
+template <class T, class DataSource, typename _DistanceType>
+struct L2_Adaptor;
+
+template <class T, class DataSource, typename _DistanceType>
+struct L1_Adaptor;
+
+template <typename Distance, class DatasetAdaptor, int DIM, typename IndexType>
+class KDTreeSingleIndexAdaptor;
+
+struct SearchParams;
+};  // namespace nanoflann
 
 namespace open3d {
 namespace core {
@@ -46,7 +62,7 @@ HOST_DEVICE inline size_t SpatialHash(int x, int y, int z) {
     return x * 73856096 ^ y * 193649663 ^ z * 83492791;
 }
 
-HOST_DEVICE inline size_t SpatialHash(const utility::MiniVec<int, 3>& xyz) {
+HOST_DEVICE inline size_t SpatialHash(const utility::MiniVec<int, 3> &xyz) {
     return SpatialHash(xyz[0], xyz[1], xyz[2]);
 }
 
@@ -57,7 +73,7 @@ HOST_DEVICE inline size_t SpatialHash(const utility::MiniVec<int, 3>& xyz) {
 ///
 template <class TVecf>
 HOST_DEVICE inline utility::MiniVec<int, 3> ComputeVoxelIndex(
-        const TVecf& pos, const typename TVecf::Scalar_t& inv_voxel_size) {
+        const TVecf &pos, const typename TVecf::Scalar_t &inv_voxel_size) {
     TVecf ref_coord = pos * inv_voxel_size;
 
     utility::MiniVec<int, 3> voxel_index;
@@ -65,6 +81,73 @@ HOST_DEVICE inline utility::MiniVec<int, 3> ComputeVoxelIndex(
     return voxel_index;
 }
 #undef HOST_DEVICE
+
+/// Base struct for Index holder
+struct NanoFlannIndexHolderBase {
+    virtual ~NanoFlannIndexHolderBase() {}
+};
+
+/// NanoFlann Index Holder.
+template <int METRIC, class TReal, class TIndex>
+struct NanoFlannIndexHolder : NanoFlannIndexHolderBase {
+    /// This class is the Adaptor for connecting Open3D Tensor and NanoFlann.
+    struct DataAdaptor {
+        DataAdaptor(size_t dataset_size,
+                    int dimension,
+                    const TReal *const data_ptr)
+            : dataset_size_(dataset_size),
+              dimension_(dimension),
+              data_ptr_(data_ptr) {}
+
+        inline size_t kdtree_get_point_count() const { return dataset_size_; }
+
+        inline TReal kdtree_get_pt(const size_t idx, const size_t dim) const {
+            return data_ptr_[idx * dimension_ + dim];
+        }
+
+        template <class BBOX>
+        bool kdtree_get_bbox(BBOX &) const {
+            return false;
+        }
+
+        size_t dataset_size_ = 0;
+        int dimension_ = 0;
+        const TReal *const data_ptr_;
+    };
+
+    /// Adaptor Selector.
+    template <int M, typename fake = void>
+    struct SelectNanoflannAdaptor {};
+
+    template <typename fake>
+    struct SelectNanoflannAdaptor<L2, fake> {
+        typedef nanoflann::L2_Adaptor<TReal, DataAdaptor, TReal> adaptor_t;
+    };
+
+    template <typename fake>
+    struct SelectNanoflannAdaptor<L1, fake> {
+        typedef nanoflann::L1_Adaptor<TReal, DataAdaptor, TReal> adaptor_t;
+    };
+
+    /// typedef for KDtree.
+    typedef nanoflann::KDTreeSingleIndexAdaptor<
+            typename SelectNanoflannAdaptor<METRIC>::adaptor_t,
+            DataAdaptor,
+            -1,
+            TIndex>
+            KDTree_t;
+
+    NanoFlannIndexHolder(size_t dataset_size,
+                         int dimension,
+                         const TReal *data_ptr) {
+        adaptor_.reset(new DataAdaptor(dataset_size, dimension, data_ptr));
+        index_.reset(new KDTree_t(dimension, *adaptor_.get()));
+        index_->buildIndex();
+    }
+
+    std::unique_ptr<KDTree_t> index_;
+    std::unique_ptr<DataAdaptor> adaptor_;
+};
 
 }  // namespace nns
 }  // namespace core
